Module Collections

module Clist : sig ... end

A Clist is a 'counted' list. Essentially, its just a list with a counter that tracks its number of elements. This makes it O(1) to get its size.

module Dlist : sig ... end

module Dlist_itf : sig ... end

module Persist : sig ... end

module Treequence : sig ... end

A Treequence is an implementation `Dlist` which keeps its elements in a tree. Every operation that adds elements to Treequence is O(1). Operations that retrieve or remove elements from it will restructure the tree to expose either the front or back of the list for fast access. This allows for efficient accesses on both ends but there is a cost to be paid when switching between accessing from the front/back or the back.

module Util : sig ... end