Module Knights_tour

module Cache : sig ... end

Disk-based memoization utility for expensive computations. Uses Marshal for serialization and a hash function for cache keys.

module Fun : sig ... end

module Game : sig ... end

module Lines : sig ... end

Utilities for reading input one line at a time. The basic idea is that different types of input which can be read one line at a time can all be converted into a string Seq.t. Then any further manipulations / operations on this input can be done in a uniform way by operating on those.

module Point : sig ... end

module PointMap : sig ... end