Module Pentominos

module Board : sig ... end

module PointSet : sig ... end

module Polyomino : sig ... end

module Puzzle : sig ... end

A polyomino puzzle consists of a board and a set of polyomino piece to be placed onto the board. The puzzle is considered solved when all pieces have been placed.