knights_tour index

Library knights_tour

The entry point of this library is the module: Knights_tour.

Library knights_tour.collections

The entry point of this library is the module: Collections.

Library knights_tour.pentominos

The entry point of this library is the module: Pentominos.

Library knights_tour.searchspace

This library exposes the following toplevel modules:

Searchspace

Stochastic_estimator Interface for stochastic estimation functions