Module Pentominos

module Board : sig ... end

A Polyomino board is a grid of squares that need to be covered by Polyomino puzzle pieces. It is possible for some parts of the grid to be blocked off.

module PointSet : sig ... end

module Polyomino : sig ... end

module Puzzle : sig ... end

A polyomino puzzle consists of a board and a set of polyomino pieces to be placed onto the board. The puzzle is considered solved when all pieces have been placed.

module Randomize : sig ... end

Utility functions for randomizing things.