Module Pentominos.Puzzle

A polyomino puzzle consists of a board and a set of polyomino piece to be placed onto the board. The puzzle is considered solved when all pieces have been placed.

type t = {

pieces : Polyomino.t list;	(*
board upon which to place the pieces
*)

board : Board.t;

}

Data type representing the state of a puzzle. This consists of a set of pieces to be placed onto the board

val classic : t

The initial state of the 'classic' Pentominos puzzle.

val solve : t -> Board.t Knights_tour.Searchspace.t